Randomized swapping algorithm
This algorithm will conduct 5000 iterations/swaps on an initial schedule and will document the DR at each swap. If a swap results in a DR improvement, the changed schedule is kept and subsequent swaps are based on the changed schedule. If a swap does not improve the DR, the schedule is not changed